Reduce redundant webview organize paths
