+++ packages/extension/src/webview/components/graph/interactionRuntime/handlers/view/zoom.ts
@@ -1,5 +1,5 @@
 import type { GraphInteractionHandlersDependencies } from '../../handlers';
-import type { GraphView2dControls } from '../../fit/api';
+import type { GraphView2dControls } from '../../fit/api/controls';
 
 export function zoom2d(
   dependencies: GraphInteractionHandlersDependencies,
