+++ packages/extension/src/webview/export/svg/node/image/overlay.ts
