+++ packages/extension/src/webview/export/svg/shape/regularPolygon/path.ts
