+++ packages/extension/src/webview/theme/brightness/detect.ts
