+++ packages/extension/tests/webview/export/svg/link/element/build.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { DEFAULT_DIRECTION_COLOR } from '../../../../../../src/shared/fileColors';
-import { buildLinkElement } from '../../../../../../src/webview/export/svg/link/element/element';
+import { buildLinkElement } from '../../../../../../src/webview/export/svg/link/element/build';
 import type { SvgExportLink } from '../../../../../../src/webview/export/svg/contracts';
 
 describe('exportSvgLinkElement', () => {
